Refactor command imports to streamline access from a centralized index; improve code organization and maintainability.
