+++ pvm/ts/src/commands/accountCommand.ts
@@ -1,6 +1,6 @@
 import { getMempoolInstance, Mempool } from "../core/mempool";
 import { Account } from "../models";
-import { getAccountManagementInstance } from "../state/accountManagement";
+import { getAccountManagementInstance } from "../state/index";
 import { IAccountManagement } from "../state/interfaces";
 import { signResult } from "./abstractSignedCommand";
 import { ISignedCommand, ISignedResponse } from "./interfaces";
