+++ pvm/ts/src/commands/createAccountCommand.ts
@@ -1,5 +1,5 @@
 import { Account } from "../models/account";
-import { getAccountManagementInstance } from "../state/accountManagement";
+import { getAccountManagementInstance } from "../state/index";
 import { IAccountManagement } from "../state/interfaces";
 import { signResult } from "./abstractSignedCommand";
 import { ISignedCommand, ISignedResponse } from "./interfaces";
