+++ pvm/ts/src/commands/getBlocksCommand.ts
@@ -1,9 +1,9 @@
 
 import { Block } from "../models/block";
-import { BlockchainManagement, getBlockchainInstance } from "../state/blockchainManagement";
+import { getBlockchainInstance } from "../state/index";
 import { signResult } from "./abstractSignedCommand";
 import { ISignedCommand, ISignedResponse } from "./interfaces";
-import { getTransactionInstance, TransactionManagement } from "../state/transactionManagement";
+import { getTransactionInstance } from "../state/index";
 import { IBlockchainManagement, ITransactionManagement } from "../state/interfaces";
 
 export class GetBlocksCommand implements ISignedCommand<Block[]> {
