+++ pvm/ts/src/commands/getTransactionsCommand.ts
@@ -1,6 +1,6 @@
 import { Transaction } from "../models";
 import { ITransactionManagement } from "../state/interfaces";
-import { getTransactionInstance } from "../state/transactionManagement";
+import { getTransactionInstance } from "../state/index";
 import { signResult } from "./abstractSignedCommand";
 import { ISignedCommand, ISignedResponse } from "./interfaces";
